Fix catkin build issues
