+++ open3d_slam/src/TransformInterpolationBuffer.cpp
@@ -4,10 +4,11 @@
  *  Created on: Nov 9, 2021
  *      Author: jelavice
  */
-
 #include "open3d_slam/TransformInterpolationBuffer.hpp"
 #include "open3d_slam/time.hpp"
 #include "open3d_slam/assert.hpp"
+
+#include <iostream>
 
 namespace o3d_slam {
